+++ src/index.js
@@ -3,4 +3,4 @@ export { default as createAction } from './action-helpers/create-action'
 export { default as createActions } from './action-helpers/create-actions'
 export { default as createReducer } from './reducer-helpers/create-reducer'
 export { default as createSelectors } from './selector-helpers/create-selectors'
-export { default as mount } from './mount'
+export { default as createLogic } from './create-logic'
