+++ test/index.test.js
@@ -3,7 +3,7 @@
 import * as reduxModular from '../src/index'
 
 it('exports modularize and createReducer', () => {
-  expect(reduxModular).toHaveProperty('mount')
+  expect(reduxModular).toHaveProperty('createLogic')
   expect(reduxModular).toHaveProperty('createType')
   expect(reduxModular).toHaveProperty('createAction')
   expect(reduxModular).toHaveProperty('createActions')
